+++ CHANGELOG.md
@@ -7,6 +7,7 @@ A more detailed list of changes is available in the corresponding milestones for
   - **[com.google.fonts/check/monospace_max_advancewidth]**: (issue #2749)
 
 ### Changes to existing checks
+  - **[[com.google.fonts/check/metadata/valid_copyright]]**: Add rationale and make it case insensitive (issue #2736)
   - **[com.google.fonts/check/metadata/undeclared_fonts]**: Clarify rationale (issue #2751)
   - **[com.google.fonts/check/metadata/filenames]**: Add rationale (issue #2751)
   - **[com.google.fonts/check/metadata/filenames]**: Consider all files from a directory (issue #2751)